requirements.txt
easyocr
presidio-analyzer
pdf2image
opencv-python
Pillow
numpy
torch
torchvision
torchaudio
